fix: apply same logic to ScreenContainer
