+++ android/src/main/java/com/swmansion/rnscreens/ScreensShadowNode.kt
@@ -11,7 +11,7 @@ internal class ScreensShadowNode(private var mContext: ReactContext) : LayoutSha
         super.onBeforeLayout(nativeViewHierarchyOptimizer)
         (mContext.getNativeModule(UIManagerModule::class.java))?.addUIBlock { nativeViewHierarchyManager: NativeViewHierarchyManager ->
             val view = nativeViewHierarchyManager.resolveView(reactTag)
-            if (view is ScreenStack) {
+            if (view is ScreenContainer<*>) {
                 view.performUpdates()
             }
         }
